add navbar in the rest of the app and remove older navbar
