--- src/components/UI/Navbar/components/CloseIcon/index.tsx
@@ -1,20 +1,0 @@
-import React from "react";
-
-function Icon() {
-  return (
-    <svg
-      xmlns="http://www.w3.org/2000/svg"
-      width="14"
-      height="14"
-      fill="none"
-      viewBox="0 0 14 14"
-    >
-      <path
-        fill="#52752C"
-        d="M13.3.71a.996.996 0 00-1.41 0L7 5.59 2.11.7A.996.996 0 10.7 2.11L5.59 7 .7 11.89a.996.996 0 101.41 1.41L7 8.41l4.89 4.89a.996.996 0 101.41-1.41L8.41 7l4.89-4.89c.38-.38.38-1.02 0-1.4z"
-      ></path>
-    </svg>
-  );
-}
-
-export default Icon;
